refactor : meta,, domain 데이터 베이스 분리 코드 수정
